Resolve unbounded process
1. Run in a process so the entry is removed from the Tempfile process afterwards.
2. Manually remove the file since Tempfile.random is only used to generate a partial path
then the extension is appended to it so the path writen to is different than the path
created by Tempfile.

Yes this is not the ideal solution, as it also breaks the tests due to the meck package
used for mocking. But it fixes the flooding of tmp directory for now and avoids an
unbounded process.